Rename composables from snake_case to camelCase
For some reason, I want to do this. We don't have a strongly consistent
filename casing convention in our `app/javascript/` directory, and this
makes that issue somewhat worse. But I still want to try it. I think
this feels right to me.
