+++ app/javascript/lib/composables/useExternalLinkTracking.ts
