+++ app/javascript/lib/composables/useFuzzyTypeahead.ts
