+++ app/javascript/lib/composables/useScrollTracking.ts
